Make some warnings about bad C-code

+++ OMCompiler/SimulationRuntime/c/simulation/solver/nonlinearSolverHomotopy.c
@@ -75,8 +75,8 @@ typedef struct DATA_HOMOTOPY
 {
   int initialized; /* 1 = initialized, else = 0*/
 
-  int n; /* dimension; n == size */
-  int m; /* dimension: m == size+1 */
+  size_t n; /* dimension; n == size */
+  size_t m; /* dimension: m == size+1 */
 
   double xtol_sqrd; /* tolerance for updating solution vector */
   double ftol_sqrd; /* tolerance for accepting accuracy */
@@ -163,7 +163,7 @@ typedef struct DATA_HOMOTOPY
  *  allocate memory for nonlinear system solver
  *  \author bbachmann
  */
-int allocateHomotopyData(int size, void** voiddata)
+int allocateHomotopyData(size_t size, void** voiddata)
 {
   DATA_HOMOTOPY* data = (DATA_HOMOTOPY*) malloc(sizeof(DATA_HOMOTOPY));
 
@@ -681,6 +681,7 @@ double vecScalarProd(int n, double *a, double *b)
 void matVecMult(int n, int m, double *A, double *b, double *c)
 {
   int i, j;
+  #warning "Fix the performance problem here; this order will trash the cache"
   for (i=0;i<n;i++) {
     c[i] = 0.0;
     for (j=0;j<m;j++)
@@ -692,6 +693,7 @@ void matVecMult(int n, int m, double *A, double *b, double *c)
 void matVecMultAbs(int n, int m, double *A, double *b, double *c)
 {
   int i, j;
+  #warning "Fix the performance problem here; this order will trash the cache"
   for (i=0;i<n;i++) {
     c[i] = 0.0;
     for (j=0;j<m;j++)
@@ -703,6 +705,7 @@ void matVecMultAbs(int n, int m, double *A, double *b, double *c)
 void matVecMultBB(int n, double *A, double *b, double *c)
 {
   int i, j;
+  #warning "Fix the performance problem here; this order will trash the cache"
   for (i=0;i<n;i++) {
     c[i] = 0.0;
     for (j=0;j<n;j++)
@@ -714,6 +717,7 @@ void matVecMultBB(int n, double *A, double *b, double *c)
 void matVecMultAbsBB(int n, double *A, double *b, double *c)
 {
   int i, j;
+  #warning "Fix the performance problem here; this order will trash the cache"
   for (i=0;i<n;i++) {
     c[i] = 0.0;
     for (j=0;j<n;j++)
@@ -725,7 +729,7 @@ void matVecMultAbsBB(int n, double *A, double *b, double *c)
 void matAddBB(int n, double* A, double* B, double* C)
 {
   int i, j;
-
+  #warning "Fix the performance problem here; this order will trash the cache"
   for (i=0;i<n;i++) {
     for (j=0;j<n+1;j++)
       C[i + j*n] = A[i + j*n] + B[i + j*n];
@@ -736,7 +740,7 @@ void matAddBB(int n, double* A, double* B, double* C)
 void matDiffBB(int n, double* A, double* B, double* C)
 {
   int i, j;
-
+  #warning "Fix the performance problem here; this order will trash the cache"
   for (i=0;i<n;i++) {
     for (j=0;j<n;j++)
       C[i + j*n] = A[i + j*n] - B[i + j*n];
@@ -749,6 +753,7 @@ void scaleMatrixRows(int n, int m, double *A)
   const double delta = sqrt(DBL_EPSILON);
   int i, j;
   double rowMax;
+  #warning "Fix the performance problem here; this order will trash the cache"
   for (i=0;i<n;i++) {
     rowMax = 0; /* This might be changed to delta */
     for (j=0;j<n;j++) {
@@ -767,6 +772,7 @@ void scaleMatrixRows(int n, int m, double *A)
 void orthogonalBacktraceMatrix(DATA_HOMOTOPY* solverData, double* hJac, double* hvec, double* v, double* hJac2, int n, int m)
 {
   int i, j;
+  #warning "Fix the performance problem here; this order will trash the cache"
   for (i=0; i<n; i++)
   {
     for (j=0; j<m; j++)
